Avoid having exchange public
